requirements.txt
playwright>=1.52.0
python-dotenv>=1.2.0
aiohttp>=3.11.0
